refactor: Add PieceCreator class
Added PieceCreator class to separate factory methods.
