+++ src/main/java/pieces/Piece.java
@@ -53,66 +53,6 @@ public class Piece {
         return point;
     }
 
-    private static Piece createWhite(Type type, Position position) {
-        return new Piece(Color.WHITE, type, position);
-    }
-
-    private static Piece createBlack(Type type, Position position) {
-        return new Piece(Color.BLACK, type, position);
-    }
-
-    public static Piece createBlank(Position position) {
-        return new Piece(Color.NO_COLOR, Type.NO_PIECE, position);
-    }
-
-    public static Piece createWhitePawn(Position position) {
-        return createWhite(Type.PAWN, position);
-    }
-
-    public static Piece createBlackPawn(Position position) {
-        return createBlack(Type.PAWN, position);
-    }
-
-    public static Piece createWhiteKnight(Position position) {
-        return createWhite(Type.KNIGHT, position);
-    }
-
-    public static Piece createBlackKnight(Position position) {
-        return createBlack(Type.KNIGHT, position);
-    }
-
-    public static Piece createWhiteRook(Position position) {
-        return createWhite(Type.ROOK, position);
-    }
-
-    public static Piece createBlackRook(Position position) {
-        return createBlack(Type.ROOK, position);
-    }
-
-    public static Piece createWhiteBishop(Position position) {
-        return createWhite(Type.BISHOP, position);
-    }
-
-    public static Piece createBlackBishop(Position position) {
-        return createBlack(Type.BISHOP, position);
-    }
-
-    public static Piece createWhiteQueen(Position position) {
-        return createWhite(Type.QUEEN, position);
-    }
-
-    public static Piece createBlackQueen(Position position) {
-        return createBlack(Type.QUEEN, position);
-    }
-
-    public static Piece createWhiteKing(Position position) {
-        return createWhite(Type.KING, position);
-    }
-
-    public static Piece createBlackKing(Position position) {
-        return createBlack(Type.KING, position);
-    }
-
     public Color getColor() {
         return color;
     }
